refactor saxs organization
